+++ common/src/main/java/uk/gov/justice/services/common/converter/jackson/ObjectMapperProducer.java
@@ -4,11 +4,13 @@ import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_ABSENT;
 import static com.fasterxml.jackson.databind.SerializationFeature.WRITE_DATES_AS_TIMESTAMPS;
 import static com.fasterxml.jackson.databind.SerializationFeature.WRITE_DATES_WITH_ZONE_ID;
 import static com.fasterxml.jackson.databind.SerializationFeature.WRITE_NULL_MAP_VALUES;
+import static java.time.ZoneOffset.UTC;
+import static java.util.TimeZone.getTimeZone;
+import static uk.gov.justice.services.common.converter.ZonedDateTimes.ISO_8601;
 
 import uk.gov.justice.services.common.converter.jackson.jsr353.InclusionAwareJSR353Module;
 
 import java.text.SimpleDateFormat;
-import java.util.TimeZone;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.enterprise.inject.Produces;
@@ -24,9 +26,6 @@ import com.fasterxml.jackson.module.paramnames.ParameterNamesModule;
 @ApplicationScoped
 public class ObjectMapperProducer {
 
-    private static final String UTC = "UTC";
-    private static final String DATE_FORMAT = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'";
-
     @Produces
     public ObjectMapper objectMapper() {
         return new ObjectMapper()
@@ -37,8 +36,8 @@ public class ObjectMapperProducer {
                 .configure(WRITE_DATES_AS_TIMESTAMPS, false)
                 .configure(WRITE_DATES_WITH_ZONE_ID, false)
                 .configure(WRITE_NULL_MAP_VALUES, false)
-                .setDateFormat(new SimpleDateFormat(DATE_FORMAT))
-                .setTimeZone(TimeZone.getTimeZone(UTC))
+                .setDateFormat(new SimpleDateFormat(ISO_8601))
+                .setTimeZone(getTimeZone(UTC))
                 .setSerializationInclusion(NON_ABSENT);
     }
 }