+++ CHANGELOG.md
@@ -9,6 +9,7 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 
 - Support for generating messaging clients from RAML
 - Support for simple access control with a default implementation that allows all requests
+- Added ZonedDateTimes utility class for converting JSON strings to UTC date times
 
 ### Changed